+++ ouster_ros/src/ros.cpp
@@ -10,12 +10,9 @@
 #include <tf2/LinearMath/Transform.h>
 #include <tf2_eigen/tf2_eigen.h>
 
-#include <cassert>
 #include <chrono>
 #include <string>
 #include <vector>
-
-#include "ouster/types.h"
 
 namespace ouster_ros {